requirements.txt
apache-airflow-providers-amazon==9.2.0
apache-airflow-providers-postgres==5.11.3
boto3==1.34.90
mlflow==2.15.1
pandas==2.2.2
psycopg2-binary==2.9.9
scikit-learn==1.5.1
